+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Bump gradle to v7.2.2
 - Migrate `BloodSugarEntrySheet` to use `BaseBottomSheet`
 - Migrate `UpdatePhoneNumberDialog` to `BaseDialog`
+- Bump dagger to v2.43.2
 
 ## 2022-08-08-8359